Create Weather App Using React.js Framework
The developed application will provide real-time weather information to the user for the city they have searched. If the user searches, for the wrong city, an Error message is also displayed to user, stating that searched city is not found. We
have used API which provides with access to weather data from around the world.